Move everything out of pkg/
