+++ cmd/wol/usage.go
@@ -4,7 +4,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/sabhiram/go-wol/pkg/wol"
+	"github.com/sabhiram/go-wol/wol"
 
 	"github.com/sabhiram/go-colorize"
 )
